Split out misc functions
